+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* CLI command with limited VT100 support (just enough for the Flipper CLI).
+
 ## [1.0.0-rc.1] - 2024-09-25
 
 ### Added
@@ -15,5 +19,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * CLI sub command can be used to start a terminal session.
 * USB port detection (currently not supported on Mac OS).
 
-[Unreleased]: https://github.com/ofabel/fssdk/compare/v1.0.0-rc.1...HEAD
+[Unreleased]: https://github.com/ofabel/fssdk/compare/v1.0.0-rc.1...dev
 [1.0.0-rc.1]: https://github.com/ofabel/fssdk/releases/tag/v1.0.0-rc.1